add morpho vault helper
